% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blb_mix}
\alias{blb_mix}
\title{Mixture Model Bag of Little Bootstraps with Optional Annealing and Non-Uniform
Sampling}
\usage{
blb_mix(
  Y,
  b,
  s,
  r,
  d,
  pr_ = NULL,
  pi_ = NULL,
  mu_ = NULL,
  sd_ = NULL,
  max_iter = 10000L,
  tol = 1e-05,
  beta = 1,
  c = 1.1,
  schedule_ = NULL
)
}
\arguments{
\item{Y}{Numeric data vector}

\item{b}{Bag of Little Bootstraps sample size}

\item{s}{Number of sample (of size b) to be taken}

\item{r}{Number of Monte Carlo iterations for each sample}

\item{d}{Number of distributions in mixture model}

\item{pr_}{Optional vector of sampling probabilities. If null, sampling is
uniform}

\item{pi_}{Optional vector of prior distribution probabilities}

\item{mu_}{Optional vector of distribution means}

\item{sd_}{Optional vector of distribution standard deviations}

\item{max_iter}{Maximum number of iterations for EM algorithm. Default is
10,000}

\item{tol}{Tolerance for convergence of log likelihood for EM algorithm.}

\item{beta}{Optional parameter representing the starting \eqn{\beta} for
annealing. For the default of 1, standard EM algorithm is run. Only needs to
be specified for annealing if schedule_ is left NULL}

\item{c}{Optional secondary parameter that determines growth rate of
\eqn{\beta}. Only needs to be specified for annealing if schedule_ is left
NULL}

\item{schedule_}{Optional scheduling for annealing, if not set to NULL the
schedule given will be used in place of the beta and c parameters}
}
\value{
List containing estimates for mixture probabilities, means, and
standard deviations
}
\description{
This function returns a list containing the estimated probabilities, means,
standard deviations and log likelihoods of the fitted GMM. Implemented in
C++
}
