% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gmm}
\alias{gmm}
\title{GMM with Optional Annealing}
\usage{
gmm(
  Y,
  d,
  pi_ = NULL,
  mu_ = NULL,
  sd_ = NULL,
  max_iter = 10000L,
  tol = 1e-05,
  beta = 1,
  c = 1.1,
  schedule_ = NULL
)
}
\arguments{
\item{Y}{Numeric data vector}

\item{d}{Number of distributions in mixture model}

\item{pi_}{Optional vector of prior distribution sampling probabilities}

\item{mu_}{Optional vector of distribution means}

\item{sd_}{Optional vector of distribution standard deviations}
}
\description{
This function returns a list containing the estimated probabilities, means,
standard deviations and log likelihoods of the fitted GMM. Implemented in
C++
}
